+++ earthmover/__main__.py
@@ -4,7 +4,16 @@ import os
 import sys
 
 from earthmover.earthmover import Earthmover
+from earthmover.init import run_init
 
+# Any new command should be added to this list
+RUN = "run"
+COMPILE = "compile"
+DEPS = "deps"
+CLEAN = "clean"
+INIT = "init"
+ALLOWED_COMMANDS = [RUN, COMPILE, DEPS, CLEAN, INIT]
+command_list = ", ".join(f"`{c}`" for c in ALLOWED_COMMANDS)
 
 class ExitOnExceptionHandler(logging.StreamHandler):
     """
@@ -50,7 +59,7 @@ def main(argv=None):
     parser.add_argument('command',
         nargs="?",
         type=str,
-        help='the command to run: `run`, `compile`, `visualize`'
+        help=f'the command to run. One of: {command_list}'
         )
     parser.add_argument("-c", "--config-file",
         nargs="?",
@@ -90,9 +99,14 @@ def main(argv=None):
         action='store_true',
         help='overwrites `show_stacktrace` config in the config file to true; sets `log_level` to DEBUG'
     )
-    parser.add_argument("--results-file",
+    parser.add_argument("-r", "--results-file",
         type=str,
         help='produces a JSON output file with structured information about run results'
+    )
+    parser.add_argument("--set",
+        type=str,
+        nargs="*",
+        help='overrides a setting in the config YAML; example: --set config.tmp_dir /tmp'
     )
 
     # Set empty defaults in case they've not been populated by the user.
@@ -103,9 +117,17 @@ def main(argv=None):
     })
 
     ### Parse the user-inputs and run Earthmover, depending on the command and subcommand passed.
-    args, remaining_argv = parser.parse_known_args()
+    args, unknown_args = parser.parse_known_args()
+    if len(unknown_args) > 0:
+        unknown_args_str = ', '.join(f"`{c}`" for c in unknown_args)
+        print(f"unknown arguments {unknown_args_str} passed, use -h flag for help")
+        exit(1)
 
-    # Command: Version
+    if args.command is not None and args.command not in ALLOWED_COMMANDS:
+        print(f"unknown command '{args.command}' passed, use -h flag for help")
+        exit(1)
+
+    # -v / --version
     if args.version:
         em_dir = os.path.dirname(os.path.abspath(__file__))
         version_file = os.path.join(em_dir, 'VERSION.txt')
@@ -114,7 +136,15 @@ def main(argv=None):
             print(f"earthmover, version {VERSION}")
         exit(0)
 
-    # Command: Test
+    if args.command == INIT:
+        project_path = run_init()
+        if project_path is not None:
+            print(f"Successfully initialized new project! You can test it out by running:\n\ncd {project_path}\nearthmover run\n")
+            exit(0)
+        else:
+            exit(1)
+
+    # -t / --test
     if args.test:
         tests_dir = os.path.join( os.path.realpath(os.path.dirname(__file__)), "tests" )
         
@@ -130,7 +160,7 @@ def main(argv=None):
         em.logger.info('tests passed successfully.')
         exit(0)
 
-    ### Otherwise, initialize Earthmover for a main run.
+    ### Otherwise, initialize Earthmover to execute a command.
     if not args.config_file:
         for file in DEFAULT_CONFIG_FILES:
             test_file = os.path.join(".", file)
@@ -141,6 +171,12 @@ def main(argv=None):
 
     if not args.config_file:
         logger.error("config file not specified with `-c` flag, and no default {" + ", ".join(DEFAULT_CONFIG_FILES) + "} found")
+
+    if args.set and len(args.set)%2 != 0: # odd number of overrides
+        logger.error("overrides specified with --set must be followed by an even number of strings (key value key value ...)")
+    overrides = None
+    if args.set:
+        overrides = dict(zip(args.set[::2], args.set[1::2]))
 
     # Update state configs with those forced via the command line.
     cli_state_configs = {}
@@ -162,15 +198,16 @@ def main(argv=None):
             force=args.force,
             skip_hashing=args.skip_hashing,
             cli_state_configs=cli_state_configs,
-            results_file=args.results_file
+            results_file=args.results_file,
+            overrides=overrides,
         )
 
     except Exception as err:
         logger.exception(err, exc_info=True)
         raise  # Avoids linting error
 
-    # Subcommand: deps (parse Earthmover YAML and compile listed packages)
-    if args.command == 'deps':
+    # Command: deps (parse Earthmover YAML and compile listed packages)
+    if args.command == DEPS:
         em.logger.info(f"installing packages...")
         if args.selector != '*':
             em.logger.info("selector is ignored for package install.")
@@ -182,23 +219,35 @@ def main(argv=None):
             logger.exception(e, exc_info=em.state_configs['show_stacktrace'])
             raise
 
-    # Subcommand: compile (parse Earthmover YAML and build graph)
-    elif args.command == 'compile':
+    # Command: compile (parse Earthmover YAML and build graph)
+    elif args.command == COMPILE:
         em.logger.info(f"compiling project...")
         if args.selector != '*':
             em.logger.info("selector is ignored for compile-only run.")
 
         try:
-            em.compile()
+            em.compile(to_disk=True)
             em.logger.info("looks ok")
 
         except Exception as e:
             logger.exception(e, exc_info=em.state_configs['show_stacktrace'])
             raise
 
-    # Subcommand: run (compile + execute)
+    elif args.command == CLEAN:
+        em.logger.info(f"removing local artifacts...")
+        if args.selector != '*':
+            em.logger.info("selector is ignored for project cleaning.")
+        
+        try:
+            em.clean()
+            em.logger.info("done!")
+        except Exception as e:
+            logger.exception(e, exc_info=em.state_configs['show_stacktrace'])
+            raise
+
+    # Command: run (compile + execute)
     # This is the default if none is specified.
-    elif args.command == 'run' or not args.command:
+    elif args.command == RUN or not args.command:
         if not args.command:
             em.logger.warning("[no command specified; proceeding with `run` but we recommend explicitly giving a command]")
 
@@ -212,7 +261,7 @@ def main(argv=None):
             raise
 
     else:
-        logger.exception("unknown command, use -h flag for help")
+        logger.exception(f"unknown command '{args.command}', use -h flag for help")
         raise