+++ earthmover/yaml_parser.py
@@ -30,6 +30,33 @@ class YamlMapping(dict):
                 self[key] = val
         return self
 
+    # enables setting values inside the YamlMapping by path, e.g. `configs.set_path("config.tmp_dir", "/mypath")`
+    def set_path(self, path, value):
+        path_pieces = path.split(".")
+        current = self
+        for path_piece in path_pieces[:-1]:
+            if path_piece not in current.keys():
+                current[path_piece] = YamlMapping()
+            current = current[path_piece]
+        current[path_pieces[-1]] = self.autocast(value)
+    
+    @staticmethod
+    def autocast(value):
+        if value.lower() in ['true', 'yes', 'on', 't', 'y']:
+            return True
+        elif value.lower() in ['false', 'no', 'off', 'f', 'n']:
+            return False
+        elif '.' in value:
+            try:
+                return float(value)
+            except ValueError:
+                return value
+        else:
+            try:
+                return int(value)
+            except ValueError:
+                return value
+    
     def to_dict(self):
         """
         Convert a YAML Mapping to a standard dictionary.